refactor post
