+++ src/app/(authorized)/profile/[id]/Posts/Posts.module.scss
