+++ src/app/(authorized)/profile/[id]/Posts/Posts.tsx
@@ -1,7 +1,7 @@
 'use client';
-import Image from 'next/image';
 import s from './Posts.module.scss';
-import { ApiResponsePosts } from './types';
+import { ApiResponsePosts } from '../types';
+import { PostImg } from './Post';
 
 type Props = {
   postsData: ApiResponsePosts;
@@ -14,19 +14,7 @@ export const Posts = ({ postsData }: Props) => {
       currentPosts = i.images.filter((postImage) => postImage.width !== 640);
       return (
         <div key={i.id} className={s.imageContainer}>
-          <Image
-            // fill
-            src={
-              i.images[0]?.url
-                ? currentPosts[0].url
-                : '/img/profile/posts/post1.png'
-            }
-            alt={'post'}
-            width={234}
-            height={228}
-            key={i.id}
-            className={s.post}
-          />
+          <PostImg post={i} />
         </div>
       );
     });
