+++ src/app/(authorized)/profile/[id]/Profile.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image';
 import s from './MyProfile.module.scss';
 import { ApiResponsePosts, UserProfile } from './types';
 import { ProfileWrapper } from './ProfileWrapper';
-import { Posts } from './Posts';
-import { LoadMore } from './load-more';
+import { Posts } from './Posts/Posts';
+import { LoadMore } from './Posts/load-more';
 import { findMinId } from '@/utils/findMinId';
 
 type Props = {
